pyproject.toml
[tool.poetry]
name = "mcrlive"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {extras = ["voice"], version = "^2.3.2"}
SpeechRecognition = "^3.9.0"
pydub = "^0.25.1"
vosk = "^0.3.45"
Pillow = "^9.4.0"
opencv-python = "^4.7.0"
responses = "^0.23.1"
pandas = "^2.0.3"
nltk = "^3.8.1"
gensim = "^4.3.1"
scikit-learn = "^1.3.0"
emoji = "^2.7.0"
fasttext = "^0.9.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
